+++ source/datasources/MyButtercupDatasource.js
@@ -31,6 +31,7 @@ class MyButtercupDatasource extends TextDatasource {
         this._vaultID = vaultID; // The MyButtercup vault identifier
         this._updateID = null;
         this._createNewClient();
+        this.type = "mybuttercup";
         fireInstantiationHandlers("mybuttercup", this);
     }
 
@@ -151,7 +152,7 @@ class MyButtercupDatasource extends TextDatasource {
      * @returns {Boolean} True if it supports changing the master password
      * @memberof WebDAVDatasource
      */
-    supportsChangePassword() {
+    supportsPasswordChange() {
         return true;
     }