+++ source/index.common.js
@@ -18,7 +18,7 @@ const { registerDatasource } = require("./datasources/register.js");
 
 const { consumeGroupFacade, consumeVaultFacade, createGroupFacade, createVaultFacade } = require("./facades/vault.js");
 const { isOTPURI, isVaultFacade } = require("./facades/detection.js");
-const { consumeEntryFacade, createEntryFacade } = require("./facades/entry.js");
+const { consumeEntryFacade, createEntryFacade, fieldsToProperties } = require("./facades/entry.js");
 const { createFieldDescriptor } = require("./facades/tools.js");
 const {
     DEFAULT_ENTRY_TYPE,
@@ -42,6 +42,7 @@ const StorageInterface = require("./storage/StorageInterface.js");
 const MemoryStorageInterface = require("./storage/MemoryStorageInterface.js");
 
 const EntryFinder = require("./search/EntryFinder.js");
+const Search = require("./search/Search.js");
 const { getSharedAppEnv } = require("./env/appEnv.js");
 
 const {
@@ -100,6 +101,7 @@ module.exports = {
     MemoryStorageInterface,
 
     EntryFinder,
+    Search,
 
     consumeEntryFacade,
     consumeGroupFacade,
@@ -112,5 +114,6 @@ module.exports = {
     isVaultFacade,
 
     getSharedAppEnv,
-    getEntryURLs
+    getEntryURLs,
+    fieldsToProperties
 };