+++ source/index.web.js
@@ -1,5 +1,8 @@
 const { getSharedAppEnv } = require("./env/core/singleton.js");
 const { applyWebConfiguration } = require("./env/web/index.js");
+const LocalFileDatasource = require("./web/LocalFileDatasource.js");
+const LocalStorageInterface = require("./web/LocalStorageInterface.js");
+const localFileClient = require("./web/localFileClient.js");
 
 let __hasInitialised = false;
 
@@ -15,5 +18,10 @@ function init() {
 }
 
 module.exports = Object.assign({}, require("./index.common.js"), require("./web/index.js"), {
+    LocalFileDatasource,
+    LocalStorageInterface,
+
+    localFileClient,
+
     init
 });